HW9: rename Strategy and ObjectStreamStrategy; in FileStorage and PathStorage remove doRead and doWrite methods
