+++ src/com/javaops/webapp/storage/strategy/Strategy.java
@@ -1,4 +1,4 @@
-package com.javaops.webapp.storage;
+package com.javaops.webapp.storage.strategy;
 
 import com.javaops.webapp.model.Resume;
 
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-public interface ObjectStreamStorageInterface {
+public interface Strategy {
 
     Resume doRead(InputStream inputStream) throws IOException;
