+++ test/com/javaops/webapp/storage/PathStorageTest.java
@@ -1,8 +1,10 @@
 package com.javaops.webapp.storage;
 
+import com.javaops.webapp.storage.strategy.ObjectStreamStrategy;
+
 public class PathStorageTest extends AbstractStorageTest {
 
     public PathStorageTest() {
-        super(new PathStorage(STORAGE_DIR.getAbsolutePath(), new ObjectStreamStorage()));
+        super(new PathStorage(STORAGE_DIR.getAbsolutePath(), new ObjectStreamStrategy()));
     }
 }
